feat(#3251): number.is-nan + is-finite + is-integer

+++ eo-parser/src/main/java/org/eolang/parser/XeEoListener.java
@@ -1220,7 +1220,7 @@ public final class XeEoListener implements EoListener, Iterable<Directive> {
             data = XeEoListener.bytesToHex(
                 ByteBuffer
                     .allocate(Double.BYTES)
-                    .putDouble(((Long) Long.parseLong(text)).doubleValue())
+                    .putDouble(Double.parseDouble(text))
                     .array()
             );
         } else if (ctx.HEX() != null) {
